requirements_new.txt
fastapi==0.115.0
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.0
google-genai>=0.5.0
